requirements.txt
async-generator==1.10
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.3.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==2.0.2
Werkzeug==2.3.4
wsproto==1.2.0
